@@ -3,22 +3,16 @@
     "description": "The Simple OAuth module for Drupal",
     "type": "drupal-module",
     "require": {
-        "league/oauth2-server": "^8.0 < 8.2",
-        "lcobucci/jwt": "^3.4",
-        "steverhoades/oauth2-openid-connect-server": "^1.1",
-        "drupal/consumers": "^1.2",
-        "php": ">=7.0"
+        "lcobucci/jwt": "^4",
+        "league/oauth2-server": "^8.3",
+        "steverhoades/oauth2-openid-connect-server": "^2.4",
+        "drupal/consumers": "^1.14",
+        "php": ">=7.4"
     },
     "require-dev": {
       "phpspec/prophecy-phpunit": "^2"
     },
-    "license": "GPL-2.0+",
-    "authors": [
-        {
-            "name": "Mateu Aguiló Bosch",
-            "email": "mateu.aguilo.bosch@gmail.com"
-        }
-    ],
+    "license": "GPL-2.0-or-later",
     "extra": {
         "drush": {
             "services": {
