@@ -7,7 +7,7 @@ use Drupal\Core\File\FileSystemInterface;
 use Drupal\Core\Site\Settings;
 use League\OAuth2\Server\CryptKey;
 use League\OAuth2\Server\Repositories\AccessTokenRepositoryInterface;
-use League\OAuth2\Server\ResourceServer as LeageResourceServer;
+use League\OAuth2\Server\ResourceServer as LeagueResourceServer;
 use Symfony\Bridge\PsrHttpMessage\HttpFoundationFactoryInterface;
 use Symfony\Bridge\PsrHttpMessage\HttpMessageFactoryInterface;
 use Symfony\Component\HttpFoundation\Request;
@@ -20,30 +20,44 @@ class ResourceServer implements ResourceServerInterface {
   /**
    * The decorated resource server.
    *
-   * @var \League\OAuth2\Server\ResourceServer
+   * @var \League\OAuth2\Server\ResourceServer|null
    */
-  protected $subject;
+  protected ?LeagueResourceServer $subject;
 
   /**
    * The message factory.
    *
    * @var \Symfony\Bridge\PsrHttpMessage\HttpMessageFactoryInterface
    */
-  protected $messageFactory;
+  protected HttpMessageFactoryInterface $messageFactory;
 
   /**
    * The HTTP foundation factory.
    *
    * @var \Symfony\Bridge\PsrHttpMessage\HttpFoundationFactoryInterface
    */
-  protected $foundationFactory;
+  protected HttpFoundationFactoryInterface $foundationFactory;
 
   /**
    * The file system.
    *
    * @var \Drupal\Core\File\FileSystemInterface
    */
-  protected $fileSystem;
+  protected FileSystemInterface $fileSystem;
+
+  /**
+   * Config Factory.
+   *
+   * @var \Drupal\Core\Config\ConfigFactoryInterface
+   */
+  protected ConfigFactoryInterface $configFactory;
+
+  /**
+   * Access Token Repository.
+   *
+   * @var \League\OAuth2\Server\Repositories\AccessTokenRepositoryInterface
+   */
+  protected AccessTokenRepositoryInterface $accessTokenRepository;
 
   /**
    * ResourceServer constructor.
@@ -54,26 +68,8 @@ class ResourceServer implements ResourceServerInterface {
     HttpMessageFactoryInterface $message_factory,
     HttpFoundationFactoryInterface $foundation_factory
   ) {
-    try {
-      $public_key = $config_factory->get('simple_oauth.settings')
-        ->get('public_key');
-      $public_key_real = $this->fileSystem()->realpath($public_key);
-      if ($public_key && $public_key_real) {
-        // Initialize the crypto key, optionally disabling the permissions
-        // check.
-        $crypt_key = new CryptKey(
-          $public_key_real,
-          NULL,
-          Settings::get('simple_oauth.key_permissions_check', TRUE)
-        );
-        $this->subject = new LeageResourceServer(
-          $access_token_repository,
-          $crypt_key
-        );
-      }
-    } catch (\LogicException $exception) {
-      trigger_error($exception, E_USER_WARNING);
-    }
+    $this->accessTokenRepository = $access_token_repository;
+    $this->configFactory = $config_factory;
     $this->messageFactory = $message_factory;
     $this->foundationFactory = $foundation_factory;
   }
@@ -94,12 +90,11 @@ class ResourceServer implements ResourceServerInterface {
   /**
    * {@inheritdoc}
    */
-  public function validateAuthenticatedRequest(Request $request) {
-    if (!$this->subject) {
-      throw new \LogicException(
-        'Unable to create resource server. Make sure public and private keys are correctly configured.'
-      );
-    }
+  public function validateAuthenticatedRequest(Request $request): Request {
+    // We don't initialize the LeagueResourceServer in the constructor, due the
+    // fact that the ResourceServer is being included on every page load.
+    $this->initializeSubject();
+    assert($this->subject instanceof LeagueResourceServer);
     // Create a PSR-7 message from the request that is compatible with the OAuth
     // library.
     $psr7_request = $this->messageFactory->createRequest($request);
@@ -111,6 +106,35 @@ class ResourceServer implements ResourceServerInterface {
 
     // Convert back to the Drupal/Symfony HttpFoundation objects.
     return $this->foundationFactory->createRequest($output_psr7_request);
+  }
+
+  /**
+   * Initialize the subject (LeagueResourceServer).
+   */
+  protected function initializeSubject(): void {
+    if (!isset($this->subject)) {
+      try {
+        $public_key = $this->configFactory->get('simple_oauth.settings')
+          ->get('public_key');
+        $public_key_real = $public_key ? $this->fileSystem()->realpath($public_key) : NULL;
+        if ($public_key_real) {
+          // Initialize the crypto key, optionally disabling the permissions
+          // check.
+          $crypt_key = new CryptKey(
+            $public_key_real,
+            NULL,
+            Settings::get('simple_oauth.key_permissions_check', TRUE)
+          );
+          $this->subject = new LeagueResourceServer(
+            $this->accessTokenRepository,
+            $crypt_key
+          );
+        }
+      }
+      catch (\LogicException $exception) {
+        trigger_error($exception, E_USER_WARNING);
+      }
+    }
   }
 
 }
