@@ -19,24 +19,25 @@ class RefreshFunctionalTest extends TokenBearerFunctionalTestBase {
    *
    * @var string
    */
-  protected $refreshToken;
+  protected string $refreshToken;
 
   /**
    * {@inheritdoc}
    */
-  protected function setUp() {
+  protected function setUp(): void {
     parent::setUp();
 
     $this->scope = 'authenticated';
     $valid_payload = [
       'grant_type' => 'password',
-      'client_id' => $this->client->uuid(),
+      'client_id' => $this->client->getClientId(),
       'client_secret' => $this->clientSecret,
       'username' => $this->user->getAccountName(),
       'password' => $this->user->pass_raw,
       'scope' => $this->scope,
     ];
     $response = $this->post($this->url, $valid_payload);
+    $this->assertValidTokenResponse($response, TRUE);
     $body = (string) $response->getBody();
     $parsed_response = Json::decode($body);
     $this->refreshToken = $parsed_response['refresh_token'];
@@ -45,11 +46,11 @@ class RefreshFunctionalTest extends TokenBearerFunctionalTestBase {
   /**
    * Test the valid Refresh grant.
    */
-  public function testRefreshGrant() {
+  public function testRefreshGrant(): void {
     // 1. Test the valid response.
     $valid_payload = [
       'grant_type' => 'refresh_token',
-      'client_id' => $this->client->uuid(),
+      'client_id' => $this->client->getClientId(),
       'client_secret' => $this->clientSecret,
       'refresh_token' => $this->refreshToken,
       'scope' => $this->scope,
@@ -62,7 +63,7 @@ class RefreshFunctionalTest extends TokenBearerFunctionalTestBase {
     $parsed_response = Json::decode((string) $response->getBody());
     $valid_payload = [
       'grant_type' => 'refresh_token',
-      'client_id' => $this->client->uuid(),
+      'client_id' => $this->client->getClientId(),
       'client_secret' => $this->clientSecret,
       'refresh_token' => $parsed_response['refresh_token'],
       'scope' => $this->scope,
@@ -73,7 +74,7 @@ class RefreshFunctionalTest extends TokenBearerFunctionalTestBase {
     // 3. Test that the token token was revoked.
     $valid_payload = [
       'grant_type' => 'refresh_token',
-      'client_id' => $this->client->uuid(),
+      'client_id' => $this->client->getClientId(),
       'client_secret' => $this->clientSecret,
       'refresh_token' => $this->refreshToken,
     ];
@@ -84,83 +85,103 @@ class RefreshFunctionalTest extends TokenBearerFunctionalTestBase {
   }
 
   /**
-   * Test invalid Refresh grant.
+   * Data provider for ::testMissingRefreshGrant.
    */
-  public function testMissingRefreshGrant() {
-    $valid_payload = [
-      'grant_type' => 'refresh_token',
-      'client_id' => $this->client->uuid(),
-      'client_secret' => $this->clientSecret,
-      'refresh_token' => $this->refreshToken,
-      'scope' => $this->scope,
-    ];
-
-    $data = [
+  public function missingRefreshGrantProvider(): array {
+    return [
       'grant_type' => [
-        'error' => 'invalid_grant',
-        'code' => 400,
+        'grant_type',
+        'invalid_grant',
+        400,
       ],
       'client_id' => [
-        'error' => 'invalid_request',
-        'code' => 400,
+        'client_id',
+        'invalid_request',
+        400,
       ],
       'client_secret' => [
-        'error' => 'invalid_client',
-        'code' => 401,
+        'client_secret',
+        'invalid_client',
+        401,
       ],
       'refresh_token' => [
-        'error' => 'invalid_request',
-        'code' => 400,
+        'refresh_token',
+        'invalid_request',
+        400,
       ],
     ];
-    foreach ($data as $key => $value) {
-      $invalid_payload = $valid_payload;
-      unset($invalid_payload[$key]);
-      $response = $this->post($this->url, $invalid_payload);
-      $parsed_response = Json::decode((string) $response->getBody());
-      $this->assertSame($value['error'], $parsed_response['error'], sprintf('Correct error code %s for %s.', $value['error'], $key));
-      $this->assertSame($value['code'], $response->getStatusCode(), sprintf('Correct status code %d for %s.', $value['code'], $key));
-    }
   }
 
   /**
    * Test invalid Refresh grant.
+   *
+   * @dataProvider missingRefreshGrantProvider
    */
-  public function testInvalidRefreshGrant() {
+  public function testMissingRefreshGrant(string $key, string $error, int $code): void {
     $valid_payload = [
       'grant_type' => 'refresh_token',
-      'client_id' => $this->client->uuid(),
+      'client_id' => $this->client->getClientId(),
       'client_secret' => $this->clientSecret,
       'refresh_token' => $this->refreshToken,
       'scope' => $this->scope,
     ];
 
-    $data = [
+    $invalid_payload = $valid_payload;
+    unset($invalid_payload[$key]);
+    $response = $this->post($this->url, $invalid_payload);
+    $parsed_response = Json::decode((string) $response->getBody());
+    $this->assertSame($error, $parsed_response['error'], sprintf('Correct error code %s', $error));
+    $this->assertSame($code, $response->getStatusCode(), sprintf('Correct status code %d', $code));
+  }
+
+  /**
+   * Data provider for ::invalidRefreshProvider.
+   */
+  public function invalidRefreshProvider(): array {
+    return [
       'grant_type' => [
-        'error' => 'invalid_grant',
-        'code' => 400,
+        'grant_type',
+        'invalid_grant',
+        400,
       ],
       'client_id' => [
-        'error' => 'invalid_client',
-        'code' => 401,
+        'client_id',
+        'invalid_client',
+        401,
       ],
       'client_secret' => [
-        'error' => 'invalid_client',
-        'code' => 401,
+        'client_secret',
+        'invalid_client',
+        401,
       ],
       'refresh_token' => [
-        'error' => 'invalid_request',
-        'code' => 401,
+        'refresh_token',
+        'invalid_request',
+        401,
       ],
     ];
-    foreach ($data as $key => $value) {
-      $invalid_payload = $valid_payload;
-      $invalid_payload[$key] = $this->getRandomGenerator()->string();
-      $response = $this->post($this->url, $invalid_payload);
-      $parsed_response = Json::decode((string) $response->getBody());
-      $this->assertSame($value['error'], $parsed_response['error'], sprintf('Correct error code %s for %s.', $value['error'], $key));
-      $this->assertSame($value['code'], $response->getStatusCode(), sprintf('Correct status code %d for %s.', $value['code'], $key));
-    }
+  }
+
+  /**
+   * Test invalid Refresh grant.
+   *
+   * @dataProvider invalidRefreshProvider
+   */
+  public function testInvalidRefreshGrant(string $key, string $error, int $code): void {
+    $valid_payload = [
+      'grant_type' => 'refresh_token',
+      'client_id' => $this->client->getClientId(),
+      'client_secret' => $this->clientSecret,
+      'refresh_token' => $this->refreshToken,
+      'scope' => $this->scope,
+    ];
+
+    $invalid_payload = $valid_payload;
+    $invalid_payload[$key] = $this->randomString();
+    $response = $this->post($this->url, $invalid_payload);
+    $parsed_response = Json::decode((string) $response->getBody());
+    $this->assertSame($error, $parsed_response['error'], sprintf('Correct error code %s', $error));
+    $this->assertSame($code, $response->getStatusCode(), sprintf('Correct status code %d', $code));
   }
 
 }
