@@ -2,7 +2,7 @@
 
 namespace Drupal\simple_oauth\Entities;
 
-use Drupal\consumers\Entity\Consumer;
+use Drupal\consumers\Entity\ConsumerInterface;
 use League\OAuth2\Server\Entities\Traits\ClientTrait;
 use League\OAuth2\Server\Entities\Traits\EntityTrait;
 
@@ -11,19 +11,21 @@ class ClientEntity implements ClientEntityInterface {
   use EntityTrait, ClientTrait;
 
   /**
-   * @var \Drupal\consumers\Entity\Consumer
+   * The consumer entity.
+   *
+   * @var \Drupal\consumers\Entity\ConsumerInterface
    */
   protected $entity;
 
   /**
    * ClientEntity constructor.
    *
-   * @param \Drupal\consumers\Entity\Consumer $entity
+   * @param \Drupal\consumers\Entity\ConsumerInterface $entity
    *   The Drupal entity.
    */
-  public function __construct(Consumer $entity) {
+  public function __construct(ConsumerInterface $entity) {
     $this->entity = $entity;
-    $this->setIdentifier($entity->uuid());
+    $this->setIdentifier($entity->getClientId());
     $this->setName($entity->label());
     if ($entity->hasField('redirect')) {
       $this->redirectUri = $entity->get('redirect')->value;
