@@ -18,9 +18,16 @@ class ScopeEntity implements ScopeEntityNameInterface {
   protected $name;
 
   /**
+   * The role associated to the scope.
+   *
+   * @var \Drupal\user\RoleInterface
+   */
+  protected $role;
+
+  /**
    * Construct a ScopeEntity instance.
    *
-   * @param \Drupal\User\RoleInterface $role
+   * @param \Drupal\user\RoleInterface $role
    *   The role associated to the scope.
    */
   public function __construct(RoleInterface $role) {
@@ -33,6 +40,7 @@ class ScopeEntity implements ScopeEntityNameInterface {
   /**
    * {@inheritdoc}
    */
+  #[\ReturnTypeWillChange]
   public function jsonSerialize() {
     return $this->getIdentifier();
   }
