@@ -6,7 +6,7 @@ use Drupal\Component\Datetime\TimeInterface;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Entity\Query\QueryException;
 use Drupal\Core\Session\AccountInterface;
-use Drupal\consumers\Entity\Consumer;
+use Drupal\consumers\Entity\ConsumerInterface;
 
 /**
  * Service in charge of deleting or expiring tokens that cannot be used anymore.
@@ -61,6 +61,7 @@ class ExpiredCollector {
    */
   public function collect($limit = 0) {
     $query = $this->tokenStorage->getQuery();
+    $query->accessCheck();
     $query->condition('expire', $this->dateTime->getRequestTime(), '<');
     // If limit available.
     if (!empty($limit)) {
@@ -83,6 +84,7 @@ class ExpiredCollector {
    */
   public function collectForAccount(AccountInterface $account) {
     $query = $this->tokenStorage->getQuery();
+    $query->accessCheck();
     $query->condition('auth_user_id', $account->id());
     $query->condition('bundle', 'refresh_token', '!=');
     $entity_ids = $query->execute();
@@ -115,14 +117,15 @@ class ExpiredCollector {
   /**
    * Collect all the tokens associated a particular client.
    *
-   * @param \Drupal\consumers\Entity\Consumer $client
+   * @param \Drupal\consumers\Entity\ConsumerInterface $client
    *   The account.
    *
    * @return \Drupal\simple_oauth\Entity\Oauth2TokenInterface[]
    *   The tokens.
    */
-  public function collectForClient(Consumer $client) {
+  public function collectForClient(ConsumerInterface $client) {
     $query = $this->tokenStorage->getQuery();
+    $query->accessCheck();
     $query->condition('client', $client->id());
     if (!$entity_ids = $query->execute()) {
       return [];
