@@ -5,24 +5,63 @@ namespace Drupal\Tests\simple_oauth\Functional;
 use Drupal\Component\Serialization\Json;
 
 /**
+ * The password grant type tests.
+ *
  * @group simple_oauth
  */
 class PasswordFunctionalTest extends TokenBearerFunctionalTestBase {
 
   /**
-   * @var string
+   * Ensure public clients still validate client secrets when they are sent.
    */
-  protected $path;
+  public function testValidateSecretEvenIfPublic(): void {
+    $this->client->set('confidential', FALSE)->save();
+    // 1. Send an invalid secret.
+    $invalid_payload = [
+      'grant_type' => 'password',
+      'client_id' => $this->client->getClientId(),
+      'client_secret' => 'invalidClientSecret',
+      'username' => $this->user->getAccountName(),
+      'password' => $this->user->pass_raw,
+      'scope' => $this->scope,
+    ];
+    $response = $this->post($this->url, $invalid_payload);
+    $this->assertEquals(401, $response->getStatusCode());
+    // 2. A valid secret is validated correctly.
+    $valid_payload = [
+      'grant_type' => 'password',
+      'client_id' => $this->client->getClientId(),
+      'client_secret' => $this->clientSecret,
+      'username' => $this->user->getAccountName(),
+      'password' => $this->user->pass_raw,
+      'scope' => $this->scope,
+    ];
+    $response = $this->post($this->url, $valid_payload);
+    $this->assertValidTokenResponse($response, TRUE);
+    // 3. The client has a secret registered, so it's always required.
+    $invalid_payload = [
+      'grant_type' => 'password',
+      'client_id' => $this->client->getClientId(),
+      // No client secret sent.
+      'username' => $this->user->getAccountName(),
+      'password' => $this->user->pass_raw,
+      'scope' => $this->scope,
+    ];
+    $response = $this->post($this->url, $invalid_payload);
+    $this->assertEquals(401, $response->getStatusCode());
+  }
 
   /**
    * Test the valid Password grant.
    */
-  public function testPasswordGrant() {
+  public function testPasswordGrant(): void {
+    $this->client->set('confidential', FALSE);
+    $this->client->set('secret', NULL)->save();
     // 1. Test the valid request.
     $valid_payload = [
       'grant_type' => 'password',
-      'client_id' => $this->client->uuid(),
-      'client_secret' => $this->clientSecret,
+      'client_id' => $this->client->getClientId(),
+      // This is a public client. PKCE and secrets are tested elsewhere.
       'username' => $this->user->getAccountName(),
       'password' => $this->user->pass_raw,
       'scope' => $this->scope,
@@ -51,7 +90,7 @@ class PasswordFunctionalTest extends TokenBearerFunctionalTestBase {
     $response = $this->post($this->url, $payload_no_scope,
       [
         'auth' => [
-          $this->client->uuid(),
+          $this->client->getClientId(),
           $this->clientSecret,
         ],
       ]
@@ -60,93 +99,134 @@ class PasswordFunctionalTest extends TokenBearerFunctionalTestBase {
   }
 
   /**
-   * Test invalid Password grant.
+   * Data provider for ::testMissingPasswordGrant.
    */
-  public function testMissingPasswordGrant() {
-    $valid_payload = [
-      'grant_type' => 'password',
-      'client_id' => $this->client->uuid(),
-      'client_secret' => $this->clientSecret,
-      'username' => $this->user->getAccountName(),
-      'password' => $this->user->pass_raw,
-      'scope' => $this->scope,
-    ];
-
-    $data = [
+  public function missingPasswordGrantProvider(): array {
+    return [
       'grant_type' => [
-        'error' => 'invalid_grant',
-        'code' => 400,
+        'grant_type',
+        'invalid_grant',
+        400,
       ],
       'client_id' => [
-        'error' => 'invalid_request',
-        'code' => 400,
+        'client_id',
+        'invalid_request',
+        400,
       ],
       'client_secret' => [
-        'error' => 'invalid_client',
-        'code' => 401,
+        'client_secret',
+        'invalid_client',
+        401,
       ],
       'username' => [
-        'error' => 'invalid_request',
-        'code' => 400,
+        'username',
+        'invalid_request',
+        400,
       ],
       'password' => [
-        'error' => 'invalid_request',
-        'code' => 400,
+        'password',
+        'invalid_request',
+        400,
       ],
     ];
-    foreach ($data as $key => $value) {
-      $invalid_payload = $valid_payload;
-      unset($invalid_payload[$key]);
-      $response = $this->post($this->url, $invalid_payload);
-      $parsed_response = Json::decode((string) $response->getBody());
-      $this->assertSame($value['error'], $parsed_response['error'], sprintf('Correct error code %s for %s.', $value['error'], $key));
-      $this->assertSame($value['code'], $response->getStatusCode(), sprintf('Correct status code %d for %s.', $value['code'], $key));
-    }
   }
 
   /**
    * Test invalid Password grant.
+   *
+   * @dataProvider missingPasswordGrantProvider
    */
-  public function testInvalidPasswordGrant() {
+  public function testMissingPasswordGrant(string $key, string $error, int $code): void {
     $valid_payload = [
       'grant_type' => 'password',
-      'client_id' => $this->client->uuid(),
+      'client_id' => $this->client->getClientId(),
       'client_secret' => $this->clientSecret,
       'username' => $this->user->getAccountName(),
       'password' => $this->user->pass_raw,
       'scope' => $this->scope,
     ];
 
-    $data = [
+    $invalid_payload = $valid_payload;
+    unset($invalid_payload[$key]);
+    $response = $this->post($this->url, $invalid_payload);
+    $parsed_response = Json::decode((string) $response->getBody());
+    $this->assertSame($error, $parsed_response['error'], sprintf('Correct error code %s', $error));
+    $this->assertSame($code, $response->getStatusCode(), sprintf('Correct status code %d', $code));
+  }
+
+  /**
+   * Data provider for ::testInvalidPasswordGrant.
+   */
+  public function invalidPasswordProvider(): array {
+    return [
       'grant_type' => [
-        'error' => 'invalid_grant',
-        'code' => 400,
+        'grant_type',
+        'invalid_grant',
+        400,
       ],
       'client_id' => [
-        'error' => 'invalid_client',
-        'code' => 401,
+        'client_id',
+        'invalid_client',
+        401,
       ],
       'client_secret' => [
-        'error' => 'invalid_client',
-        'code' => 401,
+        'client_secret',
+        'invalid_client',
+        401,
       ],
       'username' => [
-        'error' => 'invalid_credentials',
-        'code' => 401,
+        'username',
+        'invalid_grant',
+        400,
       ],
       'password' => [
-        'error' => 'invalid_credentials',
-        'code' => 401,
+        'password',
+        'invalid_grant',
+        400,
       ],
     ];
-    foreach ($data as $key => $value) {
-      $invalid_payload = $valid_payload;
-      $invalid_payload[$key] = $this->getRandomGenerator()->string();
-      $response = $this->post($this->url, $invalid_payload);
-      $parsed_response = Json::decode((string) $response->getBody());
-      $this->assertSame($value['error'], $parsed_response['error'], sprintf('Correct error code %s for %s.', $value['error'], $key));
-      $this->assertSame($value['code'], $response->getStatusCode(), sprintf('Correct status code %d for %s.', $value['code'], $key));
-    }
+  }
+
+  /**
+   * Test invalid Password grant.
+   *
+   * @dataProvider invalidPasswordProvider
+   */
+  public function testInvalidPasswordGrant(string $key, string $error, int $code): void {
+    $valid_payload = [
+      'grant_type' => 'password',
+      'client_id' => $this->client->getClientId(),
+      'client_secret' => $this->clientSecret,
+      'username' => $this->user->getAccountName(),
+      'password' => $this->user->pass_raw,
+      'scope' => $this->scope,
+    ];
+
+    $invalid_payload = $valid_payload;
+    $invalid_payload[$key] = $this->randomString();
+    $response = $this->post($this->url, $invalid_payload);
+    $parsed_response = Json::decode((string) $response->getBody());
+    $this->assertSame($error, $parsed_response['error'], sprintf('Correct error code %s', $error));
+    $this->assertSame($code, $response->getStatusCode(), sprintf('Correct status code %d', $code));
+  }
+
+  /**
+   * Test invalid secret on public client.
+   */
+  public function testInvalidSecretValidatedOnPublicClient(): void {
+    $invalid_payload = [
+      'grant_type' => 'password',
+      'client_id' => $this->client->getClientId(),
+      'client_secret' => $this->randomString(),
+      'username' => $this->user->getAccountName(),
+      'password' => $this->user->pass_raw,
+      'scope' => $this->scope,
+    ];
+    // Test sending an invalid secret; this is a public client, so it's not
+    // required, but demonstrates the secret is validated when sent.
+    $this->client->set('confidential', FALSE);
+    $response = $this->post($this->url, $invalid_payload);
+    $this->assertEquals(401, $response->getStatusCode());
   }
 
 }
