@@ -4,9 +4,10 @@ namespace Drupal\Tests\simple_oauth\Functional;
 
 use Drupal\Core\Url;
 use GuzzleHttp\RequestOptions;
+use Psr\Http\Message\ResponseInterface;
 
 /**
- *
+ * Request helper trait.
  */
 trait RequestHelperTrait {
 
@@ -15,8 +16,6 @@ trait RequestHelperTrait {
    *
    * The base methods do not provide a non-form submission POST method.
    *
-   * @see https://www.drupal.org/project/drupal/issues/2908589#comment-12258839
-   *
    * @param \Drupal\Core\Url $url
    *   The URL.
    * @param array $data
@@ -24,10 +23,14 @@ trait RequestHelperTrait {
    * @param array $options
    *   Optional options to pass to client.
    *
+   * @see https://www.drupal.org/project/drupal/issues/2908589#comment-12258839
+   *
    * @return \Psr\Http\Message\ResponseInterface
    *   The response.
+   *
+   * @throws \GuzzleHttp\Exception\GuzzleException
    */
-  protected function post(Url $url, array $data, array $options = []) {
+  protected function post(Url $url, array $data, array $options = []): ResponseInterface {
     $post_url = $this->getAbsoluteUrl($url->toString());
     $session = $this->getSession();
     $session->setCookie('SIMPLETEST_USER_AGENT', drupal_generate_test_ua($this->databasePrefix));
@@ -41,11 +44,14 @@ trait RequestHelperTrait {
    * GET a resource, with options.
    *
    * @param \Drupal\Core\Url $url
+   *   The url object to perform get request on.
    * @param array $options
+   *   The request options.
    *
-   * @return mixed
+   * @return \Psr\Http\Message\ResponseInterface
+   *   Returns the response.
    */
-  protected function get(Url $url, $options = []) {
+  protected function get(Url $url, array $options = []): ResponseInterface {
     $options += [
       RequestOptions::HTTP_ERRORS => FALSE,
     ];
