@@ -2,12 +2,13 @@
 
 namespace Drupal\Tests\simple_oauth\Unit;
 
+use Prophecy\PhpUnit\ProphecyTrait;
 use Drupal\Component\Datetime\TimeInterface;
 use Drupal\Core\Entity\EntityStorageInterface;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Entity\Query\QueryInterface;
 use Drupal\Core\Session\AccountInterface;
-use Drupal\consumers\Entity\Consumer;
+use Drupal\consumers\Entity\ConsumerInterface;
 use Drupal\simple_oauth\Entity\Oauth2Token;
 use Drupal\simple_oauth\ExpiredCollector;
 use Drupal\Tests\UnitTestCase;
@@ -18,11 +19,13 @@ use Drupal\Tests\UnitTestCase;
  */
 class EntityCollectorTest extends UnitTestCase {
 
+  use ProphecyTrait;
   /**
    * @covers ::collect
    */
   public function testCollect() {
-    list($expired_collector, $query,) = $this->buildProphecies();
+    list($expired_collector, $query) = $this->buildProphecies();
+    $query->accessCheck()->shouldBeCalledTimes(1);
     $query->condition('expire', 42, '<')->shouldBeCalledTimes(1);
     $this->assertEquals([1, 52], array_map(function ($entity) {
       return $entity->id();
@@ -33,9 +36,10 @@ class EntityCollectorTest extends UnitTestCase {
    * @covers ::collectForClient
    */
   public function testCollectForClient() {
-    list($expired_collector, $query,) = $this->buildProphecies();
-    $client = $this->prophesize(Consumer::class);
+    list($expired_collector, $query) = $this->buildProphecies();
+    $client = $this->prophesize(ConsumerInterface::class);
     $client->id()->willReturn(35);
+    $query->accessCheck()->shouldBeCalledTimes(1);
     $query->condition('client', 35)->shouldBeCalledTimes(1);
     $tokens = $expired_collector->collectForClient($client->reveal());
     $this->assertEquals([1, 52], array_map(function ($entity) {
@@ -50,6 +54,7 @@ class EntityCollectorTest extends UnitTestCase {
     list($expired_collector, $token_query,,, $client_storage) = $this->buildProphecies();
     $account = $this->prophesize(AccountInterface::class);
     $account->id()->willReturn(22);
+    $token_query->accessCheck()->shouldBeCalledTimes(2);
     $token_query->condition('auth_user_id', 22)->shouldBeCalledTimes(1);
     $token_query->condition('bundle', 'refresh_token', '!=')->shouldBeCalledTimes(1);
     $client_storage->loadByProperties([
@@ -82,6 +87,7 @@ class EntityCollectorTest extends UnitTestCase {
 
     $token_storage = $this->prophesize(EntityStorageInterface::class);
     $token_query = $this->prophesize(QueryInterface::class);
+    $token_query->accessCheck()->willReturn(TRUE);
     $token_query->execute()->willReturn([1 => '1', 52 => '52']);
     $token_storage->getQuery()->willReturn($token_query->reveal());
     $token1 = $this->prophesize(Oauth2Token::class);
@@ -95,9 +101,10 @@ class EntityCollectorTest extends UnitTestCase {
 
     $client_storage = $this->prophesize(EntityStorageInterface::class);
     $client_query = $this->prophesize(QueryInterface::class);
+    $client_query->accessCheck()->willReturn(TRUE);
     $client_query->execute()->willReturn([6 => '6']);
     $client_storage->getQuery()->willReturn($client_query->reveal());
-    $client6 = $this->prophesize(Consumer::class);
+    $client6 = $this->prophesize(ConsumerInterface::class);
     $client6->id()->willReturn(6);
     $client_storage->loadByProperties([
       'user_id' => 22,
